Add Synth and WavetableSynth
